+++ src/container/pages/meeting/viewMeetings/Actions/Actions.js
@@ -30,7 +30,10 @@ import CancelActions from "./CancelActions/CancelActions";
 import { _justShowDateformatBilling } from "../../../../../commen/functions/date_formater";
 import CustomPagination from "../../../../../commen/functions/customPagination/Paginations";
 import { clearAttendanceState } from "../../../../../store/actions/Attendance_Meeting";
-import { ViewToDoList } from "../../../../../store/actions/ToDoList_action";
+import {
+  ViewToDoList,
+  createTaskMeetingMQTT,
+} from "../../../../../store/actions/ToDoList_action";
 import ModalViewToDo from "../../../../todolistviewModal/ModalViewToDo";
 import { Select } from "antd";
 import {
@@ -123,6 +126,22 @@ const Actions = ({
     setStatusOptions(optionsArr);
   }, [todoStatus]);
 
+  // Remove task from mqtt response
+  useEffect(() => {
+    try {
+      if (toDoListReducer.socketTodoStatusData !== null) {
+        let payloadData = toDoListReducer.socketTodoStatusData;
+        if (payloadData.todoStatusID === 6) {
+          setActionsRows((rowsData) => {
+            return rowsData.filter((newData, index) => {
+              return newData.pK_TID !== payloadData.todoid;
+            });
+          });
+        }
+      }
+    } catch {}
+  }, [toDoListReducer.socketTodoStatusData]);
+
   useEffect(() => {
     if (removeTodo !== 0) {
       if (
@@ -139,14 +158,6 @@ const Actions = ({
     }
   }, [getTodosStatus.UpdateTodoStatusMessage, removeTodo]);
 
-  useEffect(() => {
-    if (
-      toDoListReducer.SocketTodoActivityData !== null &&
-      toDoListReducer.SocketTodoActivityData !== undefined
-    ) {
-      setActionsRows([toDoListReducer.SocketTodoActivityData, ...actionsRows]);
-    }
-  }, [toDoListReducer.SocketTodoActivityData]);
   // CHANGE HANDLER STATUS
   const statusChangeHandler = (e, statusdata) => {
     if (e === 6) {
@@ -387,16 +398,18 @@ const Actions = ({
       key: "RedCrossIcon",
       width: "50px",
       render: (text, record) => {
-        return (
-          <i>
-            <img
-              alt={"Cross"}
-              src={del}
-              className={styles["action-delete-cursor"]}
-              onClick={() => deleteActionHandler(record)}
-            />
-          </i>
-        );
+        if (Number(record?.taskCreator?.pK_UID) === Number(userID)) {
+          return (
+            <i>
+              <img
+                alt={"Cross"}
+                src={del}
+                className={styles["action-delete-cursor"]}
+                onClick={() => deleteActionHandler(record)}
+              />
+            </i>
+          );
+        }
       },
     },
   ];
@@ -420,56 +433,6 @@ const Actions = ({
       )
     );
   };
-  // for Socket Update meeting status update
-  useEffect(() => {
-    if (
-      toDoListReducer.socketTodoStatusData &&
-      Object.keys(toDoListReducer.socketTodoStatusData).length > 0
-    ) {
-      let tableRowsData = [...actionsRows];
-      var foundIndex = tableRowsData.findIndex(
-        (x) => x.pK_TID === toDoListReducer.socketTodoStatusData.todoid
-      );
-      if (foundIndex !== -1) {
-        if (Number(toDoListReducer.socketTodoStatusData.todoStatusID) === 6) {
-          let removeDeleteIndex = tableRowsData.filter(
-            (data, index) =>
-              data.pK_TID !== toDoListReducer.socketTodoStatusData.todoid
-          );
-          setActionsRows(removeDeleteIndex);
-        } else {
-          let newArr = tableRowsData.map((rowObj, index) => {
-            if (index === foundIndex) {
-              let statusID = toDoListReducer.socketTodoStatusData.todoStatusID;
-              const newData = {
-                ...rowObj,
-                status: {
-                  pK_TSID: statusID,
-                  status:
-                    statusID === 1
-                      ? "In Progress"
-                      : statusID === 2
-                      ? "Pending"
-                      : statusID === 3
-                      ? "Upcoming"
-                      : statusID === 4
-                      ? "Cancelled"
-                      : statusID === 5
-                      ? "Completed"
-                      : statusID === 6
-                      ? "Deleted"
-                      : null,
-                },
-              };
-              return newData;
-            }
-            return rowObj;
-          });
-          setActionsRows(newArr);
-        }
-      }
-    }
-  }, [toDoListReducer.socketTodoStatusData]);
 
   useEffect(() => {
     try {
@@ -486,6 +449,20 @@ const Actions = ({
       }
     } catch {}
   }, [actionMeetingReducer.todoListMeetingTask]);
+
+  useEffect(() => {
+    try {
+      if (toDoListReducer.createTaskMeeting !== null) {
+        let taskData = toDoListReducer.createTaskMeeting;
+        if (Number(taskData.meetingID) === Number(currentMeeting)) {
+          setActionsRows([...actionsRows, taskData.todoList]);
+        }
+        dispatch(createTaskMeetingMQTT(null));
+      }
+    } catch (error) {
+      console.log(error, "errorerrorerrorerrorerror");
+    }
+  }, [toDoListReducer.createTaskMeeting]);
 
   // for pagination in Create Task
   const handleForPagination = (current, pageSize) => {
